+++ packages/react-ui/components/PasswordInput/PasswordInput.tsx
@@ -21,13 +21,17 @@ import { styles } from './PasswordInput.styles';
 import { PasswordInputIcon } from './PasswordInputIcon';
 import { PasswordInputLocale, PasswordInputLocaleHelper } from './locale';
 
-export interface PasswordInputProps extends Pick<AriaAttributes, 'aria-label'>, CommonProps, InputProps {
+export interface PasswordInputProps
+  extends Pick<AriaAttributes, 'aria-label'>,
+    CommonProps,
+    Omit<InputProps, 'showClearIcon'> {
   /** Включает CapsLock детектор. */
   detectCapsLock?: boolean;
 }
 
 export interface PasswordInputState {
   visible: boolean;
+  focused: boolean;
   capsLockEnabled?: boolean | null;
 }
 
@@ -65,6 +69,7 @@ export class PasswordInput extends React.PureComponent<PasswordInputProps, Passw
 
   public state: PasswordInputState = {
     visible: false,
+    focused: false,
     capsLockEnabled: false,
   };
 
@@ -101,8 +106,8 @@ export class PasswordInput extends React.PureComponent<PasswordInputProps, Passw
         {(theme) => {
           this.theme = theme;
           return (
-            <CommonWrapper rootNodeRef={this.setRootNode} {...this.props}>
-              {this.renderMain(this.props)}
+            <CommonWrapper rootNodeRef={this.setRootNode} {...this.getProps()}>
+              {this.renderMain}
             </CommonWrapper>
           );
         }}
@@ -123,7 +128,9 @@ export class PasswordInput extends React.PureComponent<PasswordInputProps, Passw
    * @public
    */
   public blur = () => {
-    this.handleBlur();
+    if (this.input) {
+      this.input.blur();
+    }
   };
 
   private handleKeyPress = (e: React.KeyboardEvent<HTMLInputElement>) => {
@@ -162,18 +169,32 @@ export class PasswordInput extends React.PureComponent<PasswordInputProps, Passw
   };
 
   private handleToggleVisibility = () => {
-    this.setState((prevState) => ({ visible: !prevState.visible }), this.handleFocus);
+    this.setState((prevState) => ({ visible: !prevState.visible }), this.focusOnInput);
   };
 
-  private handleFocus = () => {
+  private focusOnInput = () => {
     if (this.input) {
       this.input.focus();
     }
   };
 
-  private handleBlur = () => {
-    if (this.input) {
-      this.input.blur();
+  private handleFocus = (event: React.FocusEvent<HTMLInputElement>) => {
+    if (this.state.focused) {
+      return;
+    }
+
+    this.setState({ focused: true });
+
+    if (this.props.onFocus) {
+      this.props.onFocus(event);
+    }
+  };
+
+  private handleFocusOutside = () => {
+    this.hideSymbols();
+
+    if (this.state.focused) {
+      this.setState({ focused: false });
     }
   };
 
@@ -229,10 +250,15 @@ export class PasswordInput extends React.PureComponent<PasswordInputProps, Passw
       onKeyDown: this.handleKeydown,
       onKeyPress: this.handleKeyPress,
       rightIcon: this.renderEye(),
+      onFocus: this.handleFocus,
     };
 
     return (
-      <RenderLayer onFocusOutside={this.hideSymbols} onClickOutside={this.hideSymbols}>
+      <RenderLayer
+        active={this.state.focused}
+        onFocusOutside={this.handleFocusOutside}
+        onClickOutside={this.handleFocusOutside}
+      >
         <div data-tid={PasswordInputDataTids.root} className={styles.root()}>
           <Input ref={this.refInput} type={this.state.visible ? 'text' : 'password'} {...inputProps} />
         </div>