+++ packages/react-ui/internal/CustomComboBox/CustomComboBox.tsx
@@ -3,7 +3,7 @@ import ReactDOM from 'react-dom';
 import { globalObject } from '@skbkontur/global-object';
 
 import { Nullable } from '../../typings/utility-types';
-import { Input, InputIconType } from '../../components/Input';
+import { Input, InputIconType, ShowClearIcon } from '../../components/Input';
 import { Menu } from '../Menu';
 import { InputLikeText } from '../InputLikeText';
 import { MenuItemState } from '../../components/MenuItem';
@@ -62,6 +62,7 @@ export interface CustomComboBoxProps<T>
   size?: SizeProp;
   totalCount?: number;
   value?: Nullable<T>;
+  showClearIcon?: ShowClearIcon;
   /**
    * Cостояние валидации при предупреждении.
    */
@@ -124,6 +125,7 @@ export class CustomComboBox<T> extends React.PureComponent<CustomComboBoxProps<T
   private cancelationToken: Nullable<(reason?: Error) => void> = null;
   private isMobileLayout!: boolean;
   private featureFlags!: ReactUIFeatureFlags;
+  private canOpenPopup = true;
 
   private reducer = reducer;
   public cancelLoaderDelay: () => void = () => null;
@@ -131,9 +133,13 @@ export class CustomComboBox<T> extends React.PureComponent<CustomComboBoxProps<T
   /**
    * @public
    */
-  public focus = () => {
+  public focus = (opts?: { withoutOpenDropdown?: boolean }) => {
     if (this.props.disabled) {
       return;
+    }
+
+    if (opts?.withoutOpenDropdown) {
+      this.canOpenPopup = false;
     }
 
     if (this.input) {
@@ -271,6 +277,7 @@ export class CustomComboBox<T> extends React.PureComponent<CustomComboBoxProps<T
       textValue: this.state.textValue,
       totalCount: this.props.totalCount,
       value: this.props.value,
+      showClearIcon: this.props.showClearIcon,
       warning: this.props.warning,
       'aria-describedby': this.props['aria-describedby'],
       'aria-label': this.props['aria-label'],
@@ -291,6 +298,7 @@ export class CustomComboBox<T> extends React.PureComponent<CustomComboBoxProps<T
       onInputValueChange: (value: string) => this.dispatch({ type: 'TextChange', value }),
       onInputFocus: this.handleFocus,
       onInputClick: this.handleInputClick,
+      onClearCrossClick: () => this.dispatch({ type: 'ClearCrossClick' }),
       onInputKeyDown: (event: React.KeyboardEvent) => {
         event.persist();
         this.dispatch({ type: 'KeyPress', event });
@@ -407,8 +415,13 @@ export class CustomComboBox<T> extends React.PureComponent<CustomComboBoxProps<T
     if (this.focused) {
       return;
     }
+
     this.focused = true;
-    this.dispatch({ type: 'Focus' });
+    this.dispatch({ type: 'Focus', searchOnFocus: this.canOpenPopup && this.props.searchOnFocus });
+
+    if (!this.canOpenPopup) {
+      this.canOpenPopup = true;
+    }
   };
 
   private handleMobileClose = () => {